fontconfig xserver plink mesalib xkeyboard-config git update 19 July 2016
xserver          commit 2a79be9e4dd1b6ba00b69ff40bc5257ec84d34da
xkeyboard-config commit 5e14a2fa1b695f94858bebda7e035ca5f016ca1b
fontconfig       commit 99645ff9eecbf2178199aa940703fbe8ed275867
mesa             commit 9c63224540ef0f727aae24274fa4afdf2e2376ac
putty            commit 9398d230339d5bfaa94093af89a17abf33b5dfad

+++ mesalib/src/mesa/state_tracker/st_context.h
@@ -70,7 +70,7 @@ struct u_upload_mgr;
 
 struct st_state_flags {
    GLbitfield mesa;  /**< Mask of _NEW_x flags */
-   uint64_t st;      /**< Mask of ST_NEW_x flags */
+   uint32_t st;      /**< Mask of ST_NEW_x flags */
 };
 
 struct st_tracked_state {
